refactor: extract WorkflowDiagramEdgeV2's content to make it easier to test
